Now KL_Update.bat is also an installer.
KL_Update.bat is now capable of detecting Spiral Knight's path, ergo you can run it from anywhere to update or install KnightLauncher. Unless you're using standalone and moved your installation folder from it's default installation path. That would be a problem that is as for now impossible to solve.